Separate spark functionality from engine
