+++ sourced/ml/cmd_entries/repos2bow.py
@@ -6,7 +6,7 @@ from sourced.ml.extractors import create_extractors_from_args
 from sourced.ml.models import OrderedDocumentFrequencies, QuantizationLevels
 from sourced.ml.transformers import Ignition, UastExtractor, UastDeserializer, Uast2Quant, \
     BagFeatures2DocFreq, BagFeatures2TermFreq, Uast2BagFeatures, HeadFiles, TFIDF, Cacher, \
-    Indexer, UastRow2Document, BOWWriter, Moder
+    Indexer, UastRow2Document, BOWWriter, Moder, create_parquet_loader
 from sourced.ml.utils import create_engine
 from sourced.ml.utils.engine import pipeline_graph, pause
 
@@ -17,20 +17,28 @@ def add_bow_args(my_parser: argparse.ArgumentParser):
     my_parser.add_argument(
         "--batch", default=BOWWriter.DEFAULT_CHUNK_SIZE, type=int,
         help="The maximum size of a single BOW file in bytes.")
+    my_parser.add_argument(
+        "--parquet", action="store_true", help="If it's parquet input.")
 
 
 @pause
 def repos2bow_entry_template(args, select=HeadFiles, cache_hook=None):
     log = logging.getLogger("repos2bow")
-    engine = create_engine("repos2bow-%s" % uuid4(), **args.__dict__)
     extractors = create_extractors_from_args(args)
+    session_name = "repos2bow-%s" % uuid4()
+    if args.parquet:
+        start_point = create_parquet_loader(session_name, **args.__dict__)
+        root = start_point
+    else:
+        engine = create_engine(session_name, **args.__dict__)
+        root = engine
 
-    ignition = Ignition(engine, explain=args.explain)
-    uast_extractor = ignition \
-        .link(select()) \
-        .link(UastExtractor(languages=args.languages)) \
-        .link(Moder(args.mode)) \
-        .link(Cacher.maybe(args.persist))
+        start_point = Ignition(engine, explain=args.explain) \
+            .link(select()) \
+            .link(UastExtractor(languages=args.languages))
+
+    uast_extractor = start_point.link(Moder(args.mode)).link(Cacher.maybe(args.persist))
+
     if cache_hook is not None:
         uast_extractor.link(cache_hook()).execute()
     # We link UastRow2Document after Cacher here because cache_hook() may want to have all possible
@@ -65,7 +73,7 @@ def repos2bow_entry_template(args, select=HeadFiles, cache_hook=None):
         .link(Indexer(Uast2BagFeatures.Columns.token, df_model.order)) \
         .link(BOWWriter(document_indexer, df_model, args.bow, args.batch)) \
         .execute()
-    pipeline_graph(args, log, ignition)
+    pipeline_graph(args, log, root)
 
 
 def repos2bow_entry(args):
