+++ sourced/ml/utils/__init__.py
@@ -1,4 +1,5 @@
 from sourced.ml.utils.bigartm import install_bigartm
-from sourced.ml.utils.engine import create_engine, create_spark, add_engine_args, add_spark_args, \
-    EngineConstants
+from sourced.ml.utils.spark import add_spark_args, create_spark, assemble_spark_config, \
+    SparkDefault
+from sourced.ml.utils.engine import add_engine_args, create_engine, EngineConstants
 from sourced.ml.utils.pickleable_logger import PickleableLogger
